# Fix TODO avoiding serialization in gRPC/GPU path