See #10530 and #10531 for the rationale.