@byronyi 2017-07-04
Using a [modified version](https://gist.github.com/byronyi/cb4ba0f6c5815ce78924ebcdd56989b9) of local recv_tensor [benchmark script](https://gist.github.com/yaroslavvb/e196107b5e0afc834652bd3153030c42) (thanks to @yaroslavvb), the performance improvement over current master is shown below:

Current master:
```
Local rate:       23185.76 MB/s
Distributed rate: 196.11 MB/s
```

With this patch:
```
Local rate:       26588.30 MB/s
Distributed rate: 256.12 MB/s
```