@byronyi 2017-07-05
Further investigation shows the l benchmark difference seems to come from different build options, e.g. AVX, SSE, etc. I suspect the difference is only visible after #6116 is fixed, so I'll revisit this PR then.