@byronyi 2017-08-09
This should be relevant now as we have the gRPC upgrade finished recently. With this change, it should do less work (mainly serialisation) in principle, but I have not checked.